requirements.txt
openai
requests
Pillow
rich
markdown2
torch
transformers
flask
flask-socketio
colorama
art
termcolor